nikel

+++ libft/ft_atoi.c
@@ -12,15 +12,15 @@
 
 #include "libft.h"
 
-long int	ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
-	long int	n;
-	int 		signe;
+	int	n;
+	int	signe;
 
 	n = 0;
 	signe = 1;
 	while (*str == '\t' || *str == '\n' || *str == '\v'
-			|| *str == '\f' || *str == '\r' || *str == ' ')
+		|| *str == '\f' || *str == '\r' || *str == ' ')
 		str++;
 	if (*str == '-' || *str == '+')
 	{
